break out control panel component
